pyproject.toml
[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[project]
name = "lilweirdo"
version = "0.0.1"
dependencies = [
    "discord.py",
    "python-dotenv",
    "uwuify",
    "ollama"
]

[project.optional-dependencies]
dev = ["isort", "mypy", "ruff"]

[project.scripts]
lilweirdo = "src.main:main"

[tool.hatch.build.targets.sdist]
packages = ["src/"]

[tool.hatch.build.targets.wheel]
packages = ["src/"]

[tool.isort]
skip = ["env/"]


# [tool.hatch.envs.default]
# dependencies = [
#     "isort"
# ]
# 
# [tool.hatch.envs.default.scripts]
# fmt = [
#     "isort ."
# ]
# 
